Refactor: add panic hook for tests;
If there is panic when running unittest an error will be printed:
```
panicked at 'foo', openraft/tests/life_cycle/shutdown.rs:78:13 backtrace=backtrace is disabled without --features 'bt' panic.file="openraft/tests/life_cycle/shutdown.rs" panic.line=78 panic.column=13
```

To log panic and traceback in a unittest: `cargo test --features bt`

Other changes:

- Adjust test dir layout.

- Fix: #24
