+++ openraft/tests/append_entries/main.rs
@@ -1,3 +1,5 @@
+#![cfg_attr(feature = "bt", feature(backtrace))]
+
 #[macro_use]
 #[path = "../fixtures/mod.rs"]
 mod fixtures;
@@ -10,5 +12,6 @@ mod t20_append_conflicts;
 mod t30_append_inconsistent_log;
 mod t40_append_updates_membership;
 mod t50_append_entries_with_bigger_term;
+mod t50_replication_1_voter_to_isolated_learner;
 mod t60_large_heartbeat;
 mod t90_issue_216_stale_last_log_id;
