+++ openraft/tests/append_entries/t50_replication_1_voter_to_isolated_learner.rs
@@ -2,14 +2,11 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use anyhow::Result;
-use fixtures::RaftRouter;
 use maplit::btreeset;
 use openraft::Config;
 
 use crate::fixtures::init_default_ut_tracing;
-
-#[macro_use]
-mod fixtures;
+use crate::fixtures::RaftRouter;
 
 /// Test replication to learner that is not in membership should not block.
 ///
