+++ openraft/tests/benchmark/main.rs
@@ -1,3 +1,5 @@
+#![cfg_attr(feature = "bt", feature(backtrace))]
+
 #[macro_use]
 #[path = "../fixtures/mod.rs"]
 mod fixtures;
