+++ openraft/tests/membership/main.rs
@@ -1,19 +1,20 @@
+#![cfg_attr(feature = "bt", feature(backtrace))]
+
 #[macro_use]
 #[path = "../fixtures/mod.rs"]
 mod fixtures;
 
 // The number indicate the preferred running order for these case.
 // The later tests may depend on the earlier ones.
-mod t00_learner_restart;
-mod t10_add_learner;
 
+mod t00_learner_restart;
+mod t01_singlenode;
+mod t10_add_learner;
+mod t12_concurrent_write_and_add_learner;
 mod t15_add_remove_follower;
 mod t16_change_membership_cases;
-// TODO(xp): rename it
 mod t20_change_membership;
-
 mod t25_elect_with_new_config;
-
 mod t30_commit_joint_config;
 mod t30_step_down;
 mod t40_removed_follower;
